Refactor pet profile component: move FetchPetData logic to a separate file for better organization
